# Replaced custom JsonSerializer with System.Text.Json & improved performance