--- a/src/NLog.Loki/AssemblyInfo.cs
+++ b/src/NLog.Loki/AssemblyInfo.cs
@@ -1,4 +1,5 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("NLog.Loki.Tests")]
+[assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")]
 [assembly: InternalsVisibleTo("Benchmark")]