--- a/src/NLog.Loki/HttpLokiTransport.cs
+++ b/src/NLog.Loki/HttpLokiTransport.cs
@@ -1,5 +1,7 @@
 using System.Collections.Generic;
 using System.Net.Http.Headers;
+using System.Net.Http.Json;
+using System.Text.Json;
 using System.Threading.Tasks;
 using NLog.Loki.Impl;
 using NLog.Loki.Model;
@@ -11,6 +13,14 @@ namespace NLog.Loki
     /// </remarks>
     internal class HttpLokiTransport : ILokiTransport
     {
+        private static readonly JsonSerializerOptions JsonOptions;
+
+        static HttpLokiTransport()
+        {
+            JsonOptions = new JsonSerializerOptions();
+            JsonOptions.Converters.Add(new LokiEventSerializer());
+        }
+
         private readonly LokiStreamsJsonSerializer lokiStreamsJsonSerializer = new();
         private readonly MediaTypeHeaderValue contentType = new("application/json");
         private readonly ILokiHttpClient lokiHttpClient;
@@ -29,6 +39,12 @@ namespace NLog.Loki
         {
             using var jsonStreamContent = new JsonStreamContent(contentType, lokiEvents, lokiStreamsJsonSerializer);
             _ = await lokiHttpClient.PostAsync("loki/api/v1/push", jsonStreamContent).ConfigureAwait(false);
+        }
+
+        public async Task WriteLogEventsWitNewSerializerAsync(IEnumerable<LokiEvent> lokiEvents)
+        {
+            using var content = JsonContent.Create(lokiEvents, options: JsonOptions);
+            _ = await lokiHttpClient.PostAsync("loki/api/v1/push", content).ConfigureAwait(false);
         }
     }
 }