--- a/src/NLog.Loki/HttpLokiTransport.cs
+++ b/src/NLog.Loki/HttpLokiTransport.cs
@@ -1,9 +1,7 @@
 using System.Collections.Generic;
-using System.Net.Http.Headers;
 using System.Net.Http.Json;
 using System.Text.Json;
 using System.Threading.Tasks;
-using NLog.Loki.Impl;
 using NLog.Loki.Model;
 
 namespace NLog.Loki
@@ -21,8 +19,6 @@ namespace NLog.Loki
             JsonOptions.Converters.Add(new LokiEventSerializer());
         }
 
-        private readonly LokiStreamsJsonSerializer lokiStreamsJsonSerializer = new();
-        private readonly MediaTypeHeaderValue contentType = new("application/json");
         private readonly ILokiHttpClient lokiHttpClient;
 
         public HttpLokiTransport(ILokiHttpClient lokiHttpClient)
@@ -37,14 +33,8 @@ namespace NLog.Loki
 
         public async Task WriteLogEventsAsync(IEnumerable<LokiEvent> lokiEvents)
         {
-            using var jsonStreamContent = new JsonStreamContent(contentType, lokiEvents, lokiStreamsJsonSerializer);
+            using var jsonStreamContent = JsonContent.Create(lokiEvents, options: JsonOptions);
             _ = await lokiHttpClient.PostAsync("loki/api/v1/push", jsonStreamContent).ConfigureAwait(false);
-        }
-
-        public async Task WriteLogEventsWitNewSerializerAsync(IEnumerable<LokiEvent> lokiEvents)
-        {
-            using var content = JsonContent.Create(lokiEvents, options: JsonOptions);
-            _ = await lokiHttpClient.PostAsync("loki/api/v1/push", content).ConfigureAwait(false);
         }
     }
 }