… src/NLog.Loki.Tests/Impl/JsonTextWriterExtensionsTests.cs
… src/NLog.Loki.Tests/Impl/JsonTextWriterTests.cs
… src/NLog.Loki/Impl/IJsonSerializer.cs
… src/NLog.Loki/Impl/JsonStreamContent.cs
… src/NLog.Loki/Impl/JsonTextWriter.cs
… src/NLog.Loki/Impl/JsonTextWriterExtensions.cs
… src/NLog.Loki/Impl/TypeExtensions.cs
… src/NLog.Loki/LokiStreamsJsonSerializer.cs