1. Removed custom serializer, replaced with System.Text.Json serializer for simpler code maintenance.
2. Improved performance when serializing messages, betwee, x1.8 and x4.0 in terms of execution time and allocations.


``` ini
BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19044
AMD Ryzen 7 2700, 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=6.0.100
  [Host]     : .NET Core 6.0.0 (CoreCLR 6.0.21.52210, CoreFX 6.0.21.52210), X64 RyuJIT
  DefaultJob : .NET Core 6.0.0 (CoreCLR 6.0.21.52210, CoreFX 6.0.21.52210), X64 RyuJIT
```

|When |                  Method |           Mean |         Error |          StdDev |   Gen 0 |  Gen 1 | Gen 2 |Allocated|
|-----:|----:|----:|--------:|--------:|-------:|------:|----------:|--:|
| Before |            LogException |        232.5 ns |         4.06 ns |         3.80 ns |  0.0193 | 0.0095 |     - |     121 B |
|After|             LogException |       227.0 ns |       0.98 ns |         0.87 ns |  0.0191 | 0.0095 |     - |     121 B |
| Before |   LogInfoWith2Variables |        321.0 ns |         4.98 ns |         4.66 ns |  0.0257 | 0.0129 |     - |     162 B |
|After|    LogInfoWith2Variables |       326.1 ns |       1.90 ns |         1.68 ns |  0.0253 | 0.0124 |     - |     161 B |
| Before |                 LogMany |     64,162.3 ns |     1,046.64 ns |       979.03 ns |  6.3477 | 3.1738 |     - |   40232 B |
|After|                  LogMany |    72,220.1 ns |     375.87 ns |       313.87 ns |  6.3477 | 3.1738 |     - |   40237 B |
|  Before |     WriteLogEventsAsync |  2,545,920.0 ns |   107,046.03 ns |   313,947.49 ns |       - |      - |     - |   13987 B |
|After|       WriteLogEventsAsync | 1,416,059.0 ns |  22,841.28 ns |    20,248.20 ns |       - |      - |     - |    4746 B |
|  Before | ManyWriteLogEventsAsync | 20,536,865.2 ns | 1,276,267.87 ns | 3,682,324.62 ns | 93.7500 |      - |     - |  414564 B |
|After|   ManyWriteLogEventsAsync | 9,005,055.6 ns | 531,591.39 ns | 1,525,235.34 ns | 15.6250 |      - |     - |  110967 B |